#73 refactor: common 내 폴더 정리
