+++ frontend/src/components/archiving/XButton/style.tsx
