+++ frontend/src/components/color/CheckIcon/index.tsx
