+++ frontend/src/components/color/CheckIcon/style.tsx
