+++ frontend/src/components/color/ExternalCarImage/index.tsx
@@ -2,7 +2,7 @@ import { MouseEvent, useState } from 'react';
 
 import { convertToTwoDigits } from '@/utils';
 
-import { RotateLogo } from '@/components/common/RotateLogo';
+import { RotateLogo } from '@/components/color/RotateLogo';
 import { PrevButton } from '@/components/common/PrevButton';
 import { NextButton } from '@/components/common/NextButton';
